// src/app/Models/GetFormInfoDataClass.ts
// Generated by https://quicktype.io

export class GetFormInfoDataClass {
    AutoKey:     number;
    FormCode:    string;
    FormName:    string;
    FlowTreeID:  string;
    StdNote:     string;
    CheckNote:   string;
    ViewNote:    string;
    EtcNote:     string;
    DynamicNode: string;
    CustomNode:  string;
    TableName:   string;
}